Векторы работают на интах

+++ AlgLogic/AlgorithmItnerface.cs
@@ -2,7 +2,7 @@
 {
     public interface AlgorithmItnerface
     {
-        int ExecuteAlgorithm(float[] vector);
+        int ExecuteAlgorithm(int[] vector);
     }
 
 }
